Add autocomplete component. Modify edit_profile view
autocomplete component is basically and autocomplete form. It's already
dynamin, and accept two parameters (input, array). I use this to make
an autocomplete form for city. Where user could just type the desired
city_name and the form will load the corresponding suggestion below it.

*BUG ENCOUNTER*
I suddenly encounter a bug with ci redirect(current_page) where the data
didn't get updated after i submited the form. I have to manually refresh
the page for the data to be updated on the view